Move Base64URL extensions into one place
